유틸 함수 5종 Utill.js로 분리하여 index.js 리팩토링
